+++ CHANGES.rst
@@ -9,6 +9,8 @@ Changelog
   registered for it in subplugins are called after the main AfterTransitionEvent managed in PloneMeeting
 - Fixed migration of MeetingConfig.groupsShownInDashboardFilter to MeetingConfig.groupsHiddenInDashboardFilter
 - Migrate vocabulary used for faceted criterion 'c4' (Group) to use 'Products.PloneMeeting.vocabularies.proposinggroupsforfacetedfiltervocabulary'
+- In the @@display-inserting-methods-helper-msg, make sure to only display categories/organizations is currently using it to sort items.
+  Use the already adaptable MeetingConfig.extraInsertingMethods method to manage extra inserting methods informations
 
 4.1rc1 (2019-06-11)
 -------------------